requirements.txt
absl-py==1.0.0
aiodns==3.0.0
aiohttp==3.7.0
aiohttp-socks==0.4.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bpemb==0.3.3
cachetools==5.0.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
conllu==4.4.1
cycler==0.11.0
dataclasses==0.6
Deprecated==1.2.13
elastic-transport==8.1.0
elasticsearch==8.1.1
fake-useragent==0.1.11
fastapi==0.75.0
filelock==3.6.0
flair==0.10
flatbuffers==2.0
fonttools==4.31.2
frozenlist==1.3.0
ftfy==6.1.1
gast==0.5.3
gdown==3.12.2
gensim==4.1.2
geographiclib==1.52
geopy==2.2.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletransx==2.4.2
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
httptools==0.4.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==3.10.1
Janome==0.4.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
konoha==4.6.5
langdetect==1.0.9
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
matplotlib==3.5.1
more-itertools==8.8.0
mpld3==0.3
multidict==6.0.2
multitasking==0.0.10
nest-asyncio==1.5.4
nsepy==0.8
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.8.0
segtok==1.5.11
sentencepiece==0.1.95
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
sqlitedict==2.0.0
starlette==0.17.1
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
tqdm==4.63.1
transformers==4.17.0
tweepy==4.8.0
twint @ file:///Users/apple/Desktop/codes/ion/ion-hack-back/twint
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
vaderSentiment==3.3.2
watchgod==0.8.1
wcwidth==0.2.5
websockets==10.2
Werkzeug==2.0.3
Wikipedia-API==0.5.4
wrapt==1.14.0
yarl==1.7.2
yfinance==0.1.70
zipp==3.7.0
